rename to barChartSplit update docs and create an array of scales
