+++ examples/src/index.js
@@ -1,5 +1,5 @@
 import "./index.css"
-import { BeeSwarm, TreeMap, BarChartStacked, Gantt, BarChartMultiple } from "../../src/index"
+import { BeeSwarm, TreeMap, BarChartStacked, Gantt, BarChartSplit } from "../../src/index"
 // import { BeeSwarm } from "gobierto-vizzs"
 import DOM from "./dom"
 import { mockJSON } from "./random";
@@ -16,7 +16,7 @@ async function main() {
     ["BeeSwarm", BeeSwarm, { relation: "relation", id: "title" }],
     ["Gantt", Gantt, { id: "title", y: "relation", barHeight: 15 }],
     ["BarChartStacked", BarChartStacked, { id: "title", data: mockJSON(5), x: "date", excludeColumns: ["relation","title", "group", "id", "phase", "from", "to"], orientationLegend: 'left', showLegend: true }],
-    ["BarChartMultiple", BarChartMultiple, { id: "title", data: mockJSON(1000), y: "relation", x: "group", count: "value4" }]
+    ["BarChartSplit", BarChartSplit, { id: "title", data: mockJSON(1000), y: "relation", x: "group", count: "value4" }]
 
   ].map(DOM);
 }
